feat: reorganized data subject request handlers and parsers for improved API functionality
