+++ handlers/mparticle-api/src/routers/http.ts
@@ -1,175 +1,47 @@
 import { createRoute, Router } from '@modules/routing/router';
-import type { APIGatewayProxyEvent, APIGatewayProxyResult } from 'aws-lambda';
-import { z } from 'zod';
 import type { DataSubjectRequestCallback } from '../../interfaces/data-subject-request-callback';
 import type { DataSubjectRequestForm } from '../../interfaces/data-subject-request-form';
 import type { EventBatch } from '../../interfaces/event-batch';
 import {
-	getStatusOfDataSubjectRequest,
-	processDataSubjectRequestCallback,
-	submitDataSubjectRequest,
-} from '../apis/data-subject-requests';
+	dataSubjectRequestCallbackHandler,
+	dataSubjectRequestCallbackParser,
+} from './http/data-subject-request-callback';
 import {
-	setUserAttributesForRightToErasureRequest,
-	uploadAnEventBatch,
-} from '../apis/events';
-import { validateDataSubjectRequestCallback } from '../utils/validate-data-subject-request-callback';
+	getDataSubjectRequestStatusHandler,
+	requestIdPathParser,
+} from './http/get-data-subject-request-status';
+import {
+	dataSubjectRequestFormParser,
+	submitDataSubjectRequestHandler,
+} from './http/submit-data-subject-request';
+import {
+	eventBatchParser,
+	uploadEventBatchHandler,
+} from './http/upload-event-batch';
 
 export const httpRouter = new Router([
 	createRoute<unknown, DataSubjectRequestForm>({
 		httpMethod: 'POST',
 		path: '/data-subject-requests',
-		handler: async (
-			event: APIGatewayProxyEvent,
-			parsed: { path: unknown; body: DataSubjectRequestForm },
-		): Promise<APIGatewayProxyResult> => {
-			/**
-			 * If you wish to remove users from audiences or from event forwarding during the waiting period,
-			 * set a user attribute and apply audience criteria and/or forwarding rules to exclude them.
-			 * https://docs.mparticle.com/guides/data-subject-requests/#erasure-request-waiting-period
-			 */
-			try {
-				await setUserAttributesForRightToErasureRequest(
-					parsed.body.environment,
-					parsed.body.userId,
-					parsed.body.submittedTime,
-				);
-			} catch (error) {
-				console.warn(
-					'It was not possible to set the User Attribute to remove user from audiences or from event forwarding during the waiting period.',
-					error,
-				);
-			}
-
-			return {
-				statusCode: 201,
-				body: JSON.stringify(await submitDataSubjectRequest(parsed.body)),
-			};
-		},
-		parser: {
-			body: z.object({
-				regulation: z.enum(['gdpr', 'ccpa']),
-				requestId: z.string().uuid(),
-				requestType: z.enum(['access', 'portability', 'erasure']),
-				submittedTime: z.string().datetime(),
-				userId: z.string(),
-				environment: z.enum(['production', 'development']),
-			}),
-		},
+		handler: submitDataSubjectRequestHandler(),
+		parser: dataSubjectRequestFormParser,
 	}),
 	createRoute<{ requestId: string }, unknown>({
 		httpMethod: 'GET',
 		path: '/data-subject-requests/{requestId}',
-		handler: async (
-			event: APIGatewayProxyEvent,
-			parsed: { path: { requestId: string }; body: unknown },
-		): Promise<APIGatewayProxyResult> => {
-			return {
-				statusCode: 200,
-				body: JSON.stringify(
-					await getStatusOfDataSubjectRequest(parsed.path.requestId),
-				),
-			};
-		},
-		parser: {
-			path: z.object({
-				requestId: z.string().uuid(),
-			}),
-		},
+		handler: getDataSubjectRequestStatusHandler(),
+		parser: requestIdPathParser,
 	}),
 	createRoute<{ requestId: string }, DataSubjectRequestCallback>({
 		httpMethod: 'POST',
 		path: '/data-subject-requests/{requestId}/callback',
-		handler: async (
-			event: APIGatewayProxyEvent,
-			parsed: { path: { requestId: string }; body: DataSubjectRequestCallback },
-		): Promise<APIGatewayProxyResult> => {
-			const getHeader = (key: string): string | undefined =>
-				Object.entries(event.headers).find(
-					([k]) => k.toLowerCase() === key.toLowerCase(),
-				)?.[1];
-			const callbackValidationResult = await validateDataSubjectRequestCallback(
-				getHeader('x-opendsr-processor-domain'),
-				getHeader('x-opendsr-signature'),
-				event.body,
-			);
-			if (!callbackValidationResult) {
-				return {
-					statusCode: 401,
-					body: 'Data Subject Request Callback validation failed.',
-				};
-			}
-
-			return {
-				statusCode: 202,
-				body: JSON.stringify(
-					processDataSubjectRequestCallback(parsed.path.requestId, parsed.body),
-				),
-			};
-		},
-		parser: {
-			path: z.object({
-				requestId: z.string().uuid(),
-			}),
-			body: z.object({
-				controller_id: z.string(),
-				expected_completion_time: z.string().datetime(),
-				subject_request_id: z.string().uuid(),
-				request_status: z.enum([
-					'pending',
-					'in_progress',
-					'completed',
-					'cancelled',
-				]),
-				api_version: z.string().nullable().optional(),
-				results_url: z.string().url().nullable(),
-				extensions: z
-					.record(
-						z.object({
-							domain: z.string(),
-							name: z.string(),
-							status: z.enum(['pending', 'skipped', 'sent', 'failed']),
-							status_message: z.string(),
-						}),
-					)
-					.nullable(),
-				group_id: z.string().nullable().optional(),
-			}),
-		},
+		handler: dataSubjectRequestCallbackHandler(),
+		parser: dataSubjectRequestCallbackParser,
 	}),
 	createRoute<unknown, EventBatch>({
 		httpMethod: 'POST',
 		path: '/events',
-		handler: async (
-			event: APIGatewayProxyEvent,
-			parsed: { path: unknown; body: EventBatch },
-		): Promise<APIGatewayProxyResult> => {
-			return {
-				statusCode: 201,
-				body: JSON.stringify(await uploadAnEventBatch(parsed.body)),
-			};
-		},
-		parser: {
-			body: z.object({
-				events: z.array(
-					z.object({
-						/** Custom data payload for the event */
-						data: z.record(z.string(), z.unknown()),
-
-						/** Type identifier for the event */
-						eventType: z.string(),
-					}),
-				),
-				deviceInfo: z.record(z.string(), z.unknown()),
-				userAttributes: z.record(z.string(), z.unknown()),
-				deletedUserAttributes: z.array(z.string()),
-				userIdentities: z.record(z.string(), z.unknown()),
-				applicationInfo: z.record(z.string(), z.unknown()),
-				schemaVersion: z.number(),
-				environment: z.enum(['production', 'development']),
-				context: z.record(z.string(), z.unknown()),
-				ip: z.string(),
-			}),
-		},
+		handler: uploadEventBatchHandler(),
+		parser: eventBatchParser,
 	}),
 ]);
