+++ handlers/mparticle-api/src/routers/http/submit-data-subject-request.ts
@@ -1,0 +1,46 @@
+import type { APIGatewayProxyEvent, APIGatewayProxyResult } from 'aws-lambda';
+import { z } from 'zod';
+import type { DataSubjectRequestForm } from '../../../interfaces/data-subject-request-form';
+import { submitDataSubjectRequest } from '../../apis/data-subject-requests';
+import { setUserAttributesForRightToErasureRequest } from '../../apis/events';
+
+export const dataSubjectRequestFormParser = {
+	body: z.object({
+		regulation: z.enum(['gdpr', 'ccpa']),
+		requestId: z.string().uuid(),
+		requestType: z.enum(['access', 'portability', 'erasure']),
+		submittedTime: z.string().datetime(),
+		userId: z.string(),
+		environment: z.enum(['production', 'development']),
+	}),
+};
+
+export function submitDataSubjectRequestHandler() {
+	return async (
+		event: APIGatewayProxyEvent,
+		parsed: { path: unknown; body: DataSubjectRequestForm },
+	): Promise<APIGatewayProxyResult> => {
+		/**
+		 * If you wish to remove users from audiences or from event forwarding during the waiting period,
+		 * set a user attribute and apply audience criteria and/or forwarding rules to exclude them.
+		 * https://docs.mparticle.com/guides/data-subject-requests/#erasure-request-waiting-period
+		 */
+		try {
+			await setUserAttributesForRightToErasureRequest(
+				parsed.body.environment,
+				parsed.body.userId,
+				parsed.body.submittedTime,
+			);
+		} catch (error) {
+			console.warn(
+				'It was not possible to set the User Attribute to remove user from audiences or from event forwarding during the waiting period.',
+				error,
+			);
+		}
+
+		return {
+			statusCode: 201,
+			body: JSON.stringify(await submitDataSubjectRequest(parsed.body)),
+		};
+	};
+}
